+++ src/routes/tests/array-component/AutoComplete.svelte
@@ -7,14 +7,15 @@
   import type { z } from 'zod';
   import type { ZodValidation, FormPathArrays } from '$lib';
   import type { SuperForm } from '$lib/client';
-  import { arrayProxy } from '$lib/client/proxies';
+  import { arrayProxy, type TaintOptions } from '$lib/client/proxies';
 
   export let form: SuperForm<ZodValidation<T>, unknown>;
   export let field: FormPathArrays<z.infer<T>>;
   export let options: { value: string; label: string }[];
   export let label = '';
+  export let taint: TaintOptions = true;
 
-  const { values, errors } = arrayProxy(form, field);
+  const { values, errors } = arrayProxy(form, field, { taint });
 </script>
 
 {#if label}<label for={field}>{label}</label>{/if}